requirements.txt
pandas
numpy
matplotlib
jinja2
PyYAML
fonttools
beautifulsoup4
